requirements.txt
flask
openai
twilio
python-dotenv
psycopg2